# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koperasi App API",
        "description": "RESTful API for Koperasi App mobile applications and integrations",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/master-data/cabang": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get list of branches (cabang)",
                "description": "Get list of cabang",
                "operationId": "60d95c8119ef653b9a9aaafd1c66d862",
                "responses": {
                    "200": {
                        "description": "List of branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Banjarmasin"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/master-data/unit-bisnis": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get list of business units",
                "description": "Get list of unit bisnis",
                "operationId": "7fcb120f0e8ede93e7085051dc141b16",
                "responses": {
                    "200": {
                        "description": "List of business units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Operasional"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/master-data/jabatan": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get list of positions (jabatan)",
                "description": "Get list of jabatan",
                "operationId": "b8f19e0f59af87f60e3269bd14891eab",
                "responses": {
                    "200": {
                        "description": "List of positions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "staff"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/master-data/divisi": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get list of divisions (divisi)",
                "description": "Get list of divisi",
                "operationId": "0946e9c46c1ee750b31d6acf29adc7e2",
                "responses": {
                    "200": {
                        "description": "List of divisions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Lapangan"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/master-data/department": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get list of departments",
                "description": "Get list of department",
                "operationId": "3ced4a52992c493d9ba6d9151cc846df",
                "responses": {
                    "200": {
                        "description": "List of departments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Operasional"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/master-data/tabel-tenor": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get tenor table for loan applications",
                "description": "Get tabel tenor",
                "operationId": "f30fa38de0e2d047f6390452e73d8198",
                "responses": {
                    "200": {
                        "description": "Tenor table",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nominal_min": {
                                                "type": "string",
                                                "example": "0.00"
                                            },
                                            "nominal_max": {
                                                "type": "string",
                                                "example": "1000000.00"
                                            },
                                            "tenor_maksimal_bulan": {
                                                "type": "integer",
                                                "example": 3
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/master-data/setting-simpanan": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get savings settings",
                "description": "Get setting simpanan",
                "operationId": "d31652beb8a8ffc8336a0c90c4934e36",
                "responses": {
                    "200": {
                        "description": "Savings settings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "jenis": {
                                                "type": "string",
                                                "example": "pokok"
                                            },
                                            "label": {
                                                "type": "string",
                                                "example": "Simpanan Pokok"
                                            },
                                            "nominal": {
                                                "type": "string",
                                                "example": "50000.00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}